requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Werkzeug==3.0.3
WTForms==3.1.2
SQLAlchemy>=2.0.36
email-validator==2.1.1
pandas
openpyxl
python-bidi==0.4.2
xhtml2pdf
google-auth-oauthlib
google-api-python-client
google-auth-httplib2
google-generativeai
python-dotenv
gunicorn
